+++ packages/connect-multichain/src/multichain/transports/mwp/index.ts
@@ -609,7 +609,7 @@ export class MWPTransport implements ExtendedTransport {
         this.windowFocusHandler = undefined;
       }
 
-      this.dappClient.disconnect();
+      await this.dappClient.disconnect();
     }
 
     this.notifyCallbacks({