+++ packages/connect-multichain/src/domain/multichain/index.ts
@@ -83,6 +83,8 @@ export abstract class MultichainCore extends EventEmitter<SDKEvents> {
   /**
    * Merges the given options into the current instance options.
    * Only the mergeable keys are updated (api.supportedNetworks, ui.*, mobile.*, transport.extensionId, debug).
+   * The main thing to note is that the value for `dapp` is not merged as it does not make sense for
+   * subsequent calls to `createMultichainClient` to have a different `dapp` value.
    * Used when createMultichainClient is called with an existing singleton.
    *
    * @param partial - Options to merge/overwrite onto the current instance