+++ playground/browser-playground/CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
-- Change Solana wallet standard query to Metamask instead of Metamask Connect ([#164](https://github.com/MetaMask/connect-monorepo/pull/164))
+- Change Solana Wallet Standard filter to find wallet name 'Metamask' rather than 'Metamask Connect' ([#164](https://github.com/MetaMask/connect-monorepo/pull/164))
 
 ## [0.2.0]